Fix bug in filter translation
In the filter translation the name is set to a boolean instead of the translation value.
It works for translations that are purely uppercase eg. ```KAUFEN``` because the value gets set again in Line ```348``` but does not for mixed eg. ```MIETE_PACHTEN```.

This bug can be fixed by using the ```Null Coalescing Operator```